+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Move `DocumentServerVersion` to this project
 * Add `IntegrationExampleApi#files`
 * Add `IntegrationExampleApi#upload_file`
+* Add `IntegrationExampleApi#file_data`
 
 ### Changes